moved action-dialog into module
