+++ src/SfxWeb/src/app/modules/action-dialog/DialogBodyComponent.ts
