+++ src/SfxWeb/src/app/modules/action-dialog/arm-warning/arm-warning.component.html
@@ -1,4 +1,4 @@
  <app-warning *ngIf="inputs.resourceId" description="This is an ARM managed resource. ARM amanged resources should only be modified during ARM deployments." 
         [link]="'https://ms.portal.azure.com/#@microsoft.onmicrosoft.com/resource' + inputs.resourceId + '/overview'" linkText="Any modifications done here should be replicated using ARM deployments. Click here to visit this resource in the Azure portal."
         iconSize="mif-lg"></app-warning>
-<app-action-dialog-template [inputs]="inputs"></app-action-dialog-template>
+<app-action-dialog-template [inputs]="inputs" (disableSubmit)="emitEvent($event)"></app-action-dialog-template>
