+++ src/SfxWeb/src/app/shared/shared.module.ts
@@ -10,7 +10,6 @@ import { DetailViewPartComponent } from './component/detail-view-part/detail-vie
 import { CollapseContainerComponent } from './component/collapse-container/collapse-container.component';
 import { RefreshRateComponent } from './component/refresh-rate/refresh-rate.component';
 import { DragDirective } from './directive/drag.directive';
-import { ActionDialogComponent } from './component/action-dialog/action-dialog.component';
 import { ManifestComponent } from './component/manifest/manifest.component';
 import { ActionCollectionDropDownComponent } from './component/action-collection-drop-down/action-collection-drop-down.component';
 import { NgbDropdownModule, NgbToastModule, NgbDatepickerModule, NgbTooltipModule } from '@ng-bootstrap/ng-bootstrap';
@@ -36,13 +35,10 @@ import {ClipboardModule} from '@angular/cdk/clipboard';
 import { WarningComponent } from './component/warning/warning.component';
 import { EventStoreTimelineComponent } from './component/event-store-timeline/event-store-timeline.component';
 import { FocusableDirective } from './directive/focusable.directive';
-import { ActionDialogTemplateComponent } from './component/action-dialog-template/action-dialog-template.component';
-import { DialogBodyDirective } from './component/dialog-body.directive';
-import { ArmWarningComponent } from './component/arm-warning/arm-warning.component';
 
 @NgModule({
   declarations: [NavbarComponent, ClipBoardComponent, HealthBadgeComponent, CandyBarCompactComponent, DetailViewPartComponent,
-                 CollapseContainerComponent, RefreshRateComponent, DragDirective, ActionDialogComponent, ManifestComponent,
+                 CollapseContainerComponent, RefreshRateComponent, DragDirective, ManifestComponent,
                  ActionCollectionDropDownComponent, InputComponent, ToastContainerComponent, CheckBoxComponent,
                  ClusterUpgradeBannerComponent, DualDatePickerComponent, AdvancedOptionComponent, ReplicaAddressComponent,
                  FormatDatePipe, LocalTimeComponent, DisplayTimeComponent, ToggleComponent, StatusResolverComponent,
@@ -51,7 +47,7 @@ import { ArmWarningComponent } from './component/arm-warning/arm-warning.compone
                  DisplayDurationComponent,
                  PhaseDiagramComponent,
                  NodeFilterComponent,
-                 WarningComponent, EventStoreTimelineComponent, FocusableDirective, ActionDialogTemplateComponent, DialogBodyDirective, ArmWarningComponent],
+                 WarningComponent, EventStoreTimelineComponent, FocusableDirective],
   imports: [
     CommonModule,
     RouterModule,
@@ -63,11 +59,11 @@ import { ArmWarningComponent } from './component/arm-warning/arm-warning.compone
     ClipboardModule
   ],
   exports: [NavbarComponent, ClipBoardComponent, HealthBadgeComponent, CandyBarCompactComponent, DetailViewPartComponent,
-            CollapseContainerComponent, RefreshRateComponent, DragDirective, ActionDialogComponent, ManifestComponent,
+            CollapseContainerComponent, RefreshRateComponent, DragDirective, ManifestComponent,
             ActionCollectionDropDownComponent, InputComponent, ToastContainerComponent, CheckBoxComponent,
             ClusterUpgradeBannerComponent, DualDatePickerComponent, AdvancedOptionComponent, ReplicaAddressComponent,
             FormatDatePipe, LocalTimeComponent, DisplayTimeComponent, ToggleComponent, StatusResolverComponent,
             StateInfoComponent, EssentialItemComponent, DisplayDurationComponent, PhaseDiagramComponent, NodeFilterComponent,
-            WarningComponent, EventStoreTimelineComponent, ActionDialogTemplateComponent, ArmWarningComponent]
+            WarningComponent, EventStoreTimelineComponent]
 })
 export class SharedModule { }
